requirements.txt
cryptography==43.0.1
pymongo==4.8.0
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
rich==13.9.1         
